Add alternative patterns for directory names from nzbmatrix Fix comments in daemon.php Add some debugging in daemon.php daemon.tools.php changes for PHP 5.3 daemon.tools.php improvements to PID File handling. Usernames/Passwords can now be different for different search providers, if no specific password is set, then the default is used.    $config['search']['<provider url>']['username']    $config['search']['<provider url>']['password'] manage.php now makes it more obvious what the new value of "automatic" is when changing it. modifications to some newzbin stuff not to use v3. but to use www. instead nzbmatrix_search should no longer cause breakage sometimes. search.php now remembers the current search api when pressing the search button search.php now uses correct links

+++ daemon.tools.php
@@ -166,6 +166,7 @@
 		// State Variables
 		$__daemontools['forked'] = false;
 		$__daemontools['loop'] = true;
+		$__daemontools['mypid'] = -1;
 		// usleep takes millionth of a second, not thousanth like expected.
 		$__daemontools['usleep_time'] = $looptime * 1000;
 		
@@ -189,7 +190,8 @@
 			}
 			
 			// Store the PID.
-			file_put_contents($__daemontools['pid'], getmypid());
+			$__daemontools['mypid'] = getmypid();
+			file_put_contents($__daemontools['pid'], $__daemontools['mypid']);
 		}
 		
 		// Add handler for the SIGTERM and SIGHUP signals
@@ -214,9 +216,12 @@
 			usleep($__daemontools['usleep_time']);
 		}
 		
-		// Try to delete the pid if we exited cleanly.
+		// Try to delete the pid if we exited cleanly, and its ours.
 		if (file_exists($__daemontools['pid'])) {
-			unlink($__daemontools['pid']);
+			$pidnum = file_get_contents($__daemontools['pid']);
+			if ($pidnum == $__daemontools['mypid']) {
+				unlink($__daemontools['pid']);
+			}
 		}
 		
 		// Let the application know we are exiting
